+++ src/Pinventory.Pins.Application/Errors.cs
@@ -1,6 +1,6 @@
 ﻿using FluentResults;
 
-using Pinventory.Pins.Application.Importing.Commands;
+using Pinventory.Pins.Application.Importing.Messages;
 using Pinventory.Pins.Application.Tags.Commands;
 
 namespace Pinventory.Pins.Application;
@@ -23,8 +23,8 @@ public static class Errors
 
     public static class Import
     {
-        public static Error RunningImportNotFound(CancelImportCommand command) =>
-            new NotFoundError($"Import {command.ArchiveJobId} not found for user {command.UserId}");
+        public static Error RunningImportNotFound(ICorrelatedMessage message) =>
+            new NotFoundError($"Import {message.ArchiveJobId} not found for user {message.UserId}");
 
         public class ArchiveJobExists() : Error("Archive job already exists");
     }
